ativ4-m8/requirements.txt
gradio==4.4.0
typeguard==4.1.5
langchain==0.0.336
python-dotenv==1.0.0
